test: split unit and integration suites and cover option forwarding
The two suites want opposite things from
@trpc/server/adapters/node-http: the unit tests need
nodeHTTPRequestHandler mocked, the integration tests need the real one.
Sharing a file meant mocking the module globally, stashing the real
implementation beforehand, and re-injecting it in a beforeAll - a
mock-then-unmock dance that only existed because of the shared module
registry. Separate files drop requireActual, the stashed reference and
the restore entirely.

The koa DefaultState augmentation moves to test/koa-state.ts so both
suites share one declaration.

Also adds the regression test for #23, which is what motivated that PR:
before it, the middleware forwarded a hand-picked set of options, so
onError never fired. Covered at both levels - the unit test asserts the
options reach the handler, the integration test asserts onError actually
fires on a throwing procedure. Verified both fail when option forwarding
is removed.

collectCoverageFrom keeps the report on src now that test/ has a
non-test module.
