+++ jest.config.js
@@ -6,6 +6,7 @@
 module.exports = {
   clearMocks: true,
   collectCoverage: true,
+  collectCoverageFrom: ['src/**/*.ts'],
   coverageDirectory: 'coverage',
   coverageProvider: 'v8',
   preset: 'ts-jest',
