+++ test/integration.test.ts
@@ -1,140 +1,16 @@
-import Koa, { Context } from 'koa';
-import { createKoaMiddleware, CreateTrpcKoaContextOptions } from '../src';
+import Koa from 'koa';
 import request from 'supertest';
-import { initTRPC } from '@trpc/server';
-import { nodeHTTPRequestHandler } from '@trpc/server/adapters/node-http';
+import { initTRPC, TRPCError } from '@trpc/server';
 import { Server } from 'http';
 import koaBodyParserOld from 'koa-bodyparser';
 import koaBodyParser from '@koa/bodyparser';
+import { createKoaMiddleware, CreateTrpcKoaContextOptions } from '../src';
+import './koa-state';
 
-// Augment Koa types for type-safe state access in tests
-declare module 'koa' {
-  interface DefaultState {
-    userId?: number;
-    userName?: string;
-  }
-}
-
-// Store real implementation before mocking
-const realNodeHTTPRequestHandler = jest.requireActual<
-  typeof import('@trpc/server/adapters/node-http')
->('@trpc/server/adapters/node-http').nodeHTTPRequestHandler;
-
-// Mock the module - required for tRPC v11 which has frozen exports
-jest.mock('@trpc/server/adapters/node-http', () => ({
-  ...jest.requireActual('@trpc/server/adapters/node-http'),
-  nodeHTTPRequestHandler: jest.fn(),
-}));
-
-// Get a reference to the mocked function
-const mockNodeHTTPRequestHandler = nodeHTTPRequestHandler as jest.Mock;
-
-describe('Unit', () => {
-  const router = initTRPC.create().router({});
-  const next = jest.fn();
-
-  beforeEach(() => {
-    mockNodeHTTPRequestHandler.mockReset();
-  });
-
-  it('should return a function accepting 2 arguments', () => {
-    const adapter = createKoaMiddleware({
-      router,
-    });
-    expect(typeof adapter).toBe('function');
-    expect(adapter.length).toBe(2);
-  });
-  it('createKoaMiddleware should accept 1 argument', () => {
-    expect(createKoaMiddleware.length).toBe(1);
-  });
-  it('createKoaMiddleware should call nodeHTTPRequestHandler if request prefix matches', () => {
-    const adapter = createKoaMiddleware({ router, prefix: '/trpc' });
-
-    const ctx = {
-      request: {
-        path: '/trpc/users',
-      },
-      req: {},
-      res: {},
-    } as Context;
-
-    adapter(ctx, next);
-
-    expect(next).not.toHaveBeenCalled();
-    expect(mockNodeHTTPRequestHandler).toHaveBeenCalled();
-  });
-  it('createKoaMiddleware should call nodeHTTPRequestHandler if no prefix set', () => {
-    const adapter = createKoaMiddleware({ router });
-
-    const ctx = {
-      request: {
-        path: '/users',
-      },
-      req: {},
-      res: {},
-    } as Context;
-
-    adapter(ctx, next);
-
-    expect(next).not.toHaveBeenCalled();
-    expect(mockNodeHTTPRequestHandler).toHaveBeenCalled();
-  });
-  it('createKoaMiddleware should call next and not process request if prefix set and request doesnt have prefix', () => {
-    const adapter = createKoaMiddleware({ router, prefix: '/trpc' });
-
-    const ctx = {
-      request: {
-        path: '/users', // prefix missing from path
-      },
-    } as Context;
-
-    adapter(ctx, next);
-
-    expect(next).toHaveBeenCalled();
-    expect(mockNodeHTTPRequestHandler).not.toHaveBeenCalled();
-  });
-  it('createKoaMiddleware should call nodeHTTPRequestHandler with req.body if parsed body found on request.body', () => {
-    const adapter = createKoaMiddleware({ router });
-
-    const ctx = {
-      request: { path: '/users', body: { name: 'Person1', age: 20 } },
-      req: {},
-      res: {},
-    } as Context;
-    adapter(ctx, next);
-
-    expect(next).not.toHaveBeenCalled();
-    expect(mockNodeHTTPRequestHandler).toHaveBeenCalledWith(
-      expect.objectContaining({
-        req: expect.objectContaining({ body: ctx.request.body }),
-      }),
-    );
-  });
-  it('should attach Koa context to req.koaCtx', () => {
-    const adapter = createKoaMiddleware({ router });
-    const ctx = {
-      request: { path: '/users' },
-      req: {},
-      res: {},
-      state: { userId: 123 },
-    } as Context;
-
-    adapter(ctx, next);
-
-    expect(mockNodeHTTPRequestHandler).toHaveBeenCalledWith(
-      expect.objectContaining({
-        req: expect.objectContaining({ koaCtx: ctx }),
-      }),
-    );
-  });
-});
+// This suite exercises the real nodeHTTPRequestHandler against a running Koa
+// server, so unlike the unit suite it must not mock it.
 
 describe('Integration', () => {
-  // Restore real implementation for integration tests
-  beforeAll(() => {
-    mockNodeHTTPRequestHandler.mockImplementation(realNodeHTTPRequestHandler);
-  });
-
   const ALL_USERS = [
     { id: 1, name: 'bob' },
     { id: 2, name: 'alice' },
@@ -281,7 +157,7 @@ describe('Integration', () => {
 
     const createContextWithState = ({ req }: CreateTrpcKoaContextOptions) => ({
       // TypeScript knows userId is number | undefined and userName is string | undefined
-      // due to the module augmentation above
+      // due to the module augmentation in ./koa-state
       userId: req.koaCtx?.state.userId,
       userName: req.koaCtx?.state.userName,
     });
@@ -390,6 +266,49 @@ describe('Integration', () => {
 
       expect(response.status).toEqual(200);
       expect(response.body.result.data).toEqual({ session: 'abc123' });
+    });
+  });
+
+  // The middleware used to forward only a hand-picked set of options to the
+  // handler, so onError never fired.
+  // https://github.com/BlairCurrey/trpc-koa-adapter/pull/23
+  describe('tRPC Handler Options', () => {
+    const app = new Koa();
+    const onError = jest.fn();
+
+    const trpcErrors = initTRPC.create();
+    const trpcErrorRouter = trpcErrors.router({
+      boom: trpcErrors.procedure.query(() => {
+        throw new TRPCError({ code: 'INTERNAL_SERVER_ERROR', message: 'boom' });
+      }),
+    });
+
+    app.use(
+      createKoaMiddleware({
+        router: trpcErrorRouter,
+        prefix: '/trpc',
+        onError,
+      }),
+    );
+
+    let server: Server;
+
+    beforeEach(async () => (server = app.listen(3101)));
+    afterEach(async () => await server.close());
+
+    it('should call onError when a procedure throws', async () => {
+      const response = await request(server)
+        .get('/trpc/boom')
+        .set('content-type', 'application/json');
+
+      expect(response.status).toEqual(500);
+      expect(onError).toHaveBeenCalledWith(
+        expect.objectContaining({
+          error: expect.objectContaining({ message: 'boom' }),
+          path: 'boom',
+          type: 'query',
+        }),
+      );
     });
   });
 });
